chore: handler consumer non handled events

+++ licenseware/config/config.py
@@ -89,7 +89,7 @@ class Config(BaseSettings):  # pragma no cover
     EXPIRE_MACHINE_CHECK: int = 60  # 1 minute
 
     KAFKA_BROKER_URL: str = "PLAINTEXT://localhost:9092"
-    KAFKA_CONSUMER_POLL: int = 1
+    KAFKA_CONSUMER_POLL: float = 1.0
     KAFKA_SECURITY_PROTOCOL: str = "PLAINTEXT"
 
     RABBITMQ_HOST: str = "localhost"
